+++ src/components/CheckEditor/testConstants.ts
@@ -82,6 +82,7 @@ export const BASIC_HTTP_CHECK = {
       headers: [],
       body: '',
       proxyURL: '',
+      proxyConnectHeaders: [],
       cacheBustingQueryParamName: '',
       compression: undefined,
       failIfNotSSL: false,
@@ -214,6 +215,66 @@ export const CUSTOM_ALERT_SENSITIVITY_CHECK = {
   },
 };
 
+export const BASIC_MULTIHTTP_CHECK = {
+  job: 'basicmulti',
+  target: 'https://www.grafana.com',
+  enabled: true,
+  labels: [{ name: 'labelName', value: 'labelValue' }],
+  probes: [42],
+  timeout: 2000,
+  frequency: 110000,
+  alertSensitivity: 'none',
+  basicMetricsOnly: true,
+  settings: {
+    multihttp: {
+      entries: [
+        {
+          request: {
+            url: 'https://www.grafana.com',
+            method: 'GET',
+            headers: [
+              {
+                name: 'aheader',
+                value: 'yarp',
+              },
+              {
+                name: 'carne',
+                value: 'asada',
+              },
+            ],
+            body: '',
+            queryString: [
+              {
+                name: 'tacos',
+                value: 'delicious',
+              },
+            ],
+          },
+        },
+        {
+          request: {
+            url: 'https://www.example.com',
+            method: 'POST',
+            headers: [
+              {
+                name: 'examples',
+                value: 'great',
+              },
+            ],
+            body: '{"averyinteresting":"request body content"}',
+            queryString: [
+              {
+                name: 'query',
+                value: 'param',
+              },
+            ],
+          },
+        },
+      ],
+    },
+  },
+};
+
 export const BASIC_CHECK_LIST = [
   {
     job: 'carne asada',
@@ -260,6 +321,7 @@ export const BASIC_CHECK_LIST = [
   { id: 3, ...BASIC_PING_CHECK },
   { id: 4, ...BASIC_TCP_CHECK },
   { id: 5, ...CUSTOM_ALERT_SENSITIVITY_CHECK },
+  { id: 6, ...BASIC_MULTIHTTP_CHECK },
 ] as Check[];
 
 export const EDITED_HTTP_CHECK = {
@@ -291,6 +353,7 @@ export const EDITED_HTTP_CHECK = {
         insecureSkipVerify: true,
       },
       headers: ['headerName:headerValue', 'headerName:headerValue'],
+      proxyConnectHeaders: [],
       validStatusCodes: [100],
       validHTTPVersions: ['HTTP/1.0'],
       failIfNotSSL: true,