+++ src/datasource/DataSource.ts
@@ -13,8 +13,8 @@ import {
 import { SMQuery, SMOptions, QueryType, CheckInfo, DashboardVariable } from './types';
 
 import { getBackendSrv, getTemplateSrv } from '@grafana/runtime';
-import { Probe, Check, RegistrationInfo, HostedInstance } from '../types';
-import { findLinkedDatasource, queryLogs } from 'utils';
+import { Probe, Check, HostedInstance } from '../types';
+import { findLinkedDatasource, getRandomProbes, queryLogs } from 'utils';
 import { parseTracerouteLogs } from './traceroute-utils';
 import { firstValueFrom } from 'rxjs';
 
@@ -244,6 +244,10 @@ export class SMDataSource extends DataSourceApi<SMQuery, SMOptions> {
     if (check.timeout > 2500) {
       check.timeout = 2500;
     }
+
+    const randomSelection = getRandomProbes(check.probes, 5);
+    check.probes = randomSelection;
+
     return firstValueFrom(
       getBackendSrv().fetch({
         method: 'POST',
@@ -369,33 +373,6 @@ export class SMDataSource extends DataSourceApi<SMQuery, SMOptions> {
     } else {
       return 'https://' + url;
     }
-  }
-
-  async registerInit(apiHost: string, apiToken: string): Promise<RegistrationInfo> {
-    const backendSrv = getBackendSrv();
-    const data = {
-      ...this.instanceSettings,
-      jsonData: {
-        apiHost: this.normalizeURL(apiHost),
-      },
-      access: 'proxy',
-    };
-    await backendSrv.put(`api/datasources/${this.instanceSettings.id}`, data);
-    return backendSrv
-      .fetch({
-        method: 'POST',
-        url: `${this.instanceSettings.url}/sm/register/init`,
-        data: { apiToken },
-        headers: {
-          // ensure the grafana backend doesn't use a cached copy of the
-          // datasource config, as it might not have the new apiHost set.
-          'X-Grafana-NoCache': 'true',
-        },
-      })
-      .toPromise()
-      .then((res: any) => {
-        return res.data;
-      });
   }
 
   async onOptionsChange(options: SMOptions) {